requirements.txt
fastapi
uvicorn
psycopg2-binary
python-multipart
huggingface-hub
PyPDF2
llmgraph
llmwiki
python-dotenv